+++ library/Solarium/Core/Client/Adapter/Curl.php
@@ -156,7 +156,11 @@ class Curl extends Configurable implements AdapterInterface
         }
 
         if (!isset($options['headers']['Content-Type'])) {
-            $options['headers']['Content-Type'] = 'text/xml; charset=utf-8';
+            if($method == Request::METHOD_GET){
+                $options['headers']['Content-Type'] = 'application/x-www-form-urlencoded; charset=utf-8';
+            } else {
+                $options['headers']['Content-Type'] = 'application/xml; charset=utf-8';
+            }
         }
 
         // Try endpoint authentication first, fallback to request for backwards compatibility
@@ -184,7 +188,7 @@ class Curl extends Configurable implements AdapterInterface
             if ($request->getFileUpload()) {
                 if (version_compare(PHP_VERSION, '5.5.0') >= 0) {
                     $curlFile = curl_file_create($request->getFileUpload());
-                    curl_setopt($handler, CURLOPT_POSTFIELDS, array('content', $curlFile));
+                    curl_setopt($handler, CURLOPT_POSTFIELDS, array('content' => $curlFile));
                 } else {
                     curl_setopt($handler, CURLOPT_POSTFIELDS, array('content' => '@'.$request->getFileUpload()));
                 }