+++ library/Solarium/QueryType/Select/RequestBuilder/Component/Stats.php
@@ -60,8 +60,10 @@ class Stats implements ComponentRequestBuilderInterface
 
         // add fields
         foreach ($component->getFields() as $field) {
+            $pivots = $field->getPivots();
 
-            $request->addParam('stats.field', $field->getKey());
+            $prefix = (count($pivots) > 0) ? '{!tag='.implode(',', $pivots).'}' : '';
+            $request->addParam('stats.field', $prefix . $field->getKey());
 
             // add field specific facet stats
             foreach ($field->getFacets() as $facet) {