+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+## 3.4.0 - 2015-06-14
+
+- bugfix: only check type for added documents to add query if provided as an array
+- improvement: added support for calling empty() and isset() on result document properties
+- improvement: added composer test script
+- bugfix: curl file upload handling
+- improvement: added 'contributing' file
+- improvement: docblock fixes in grouping component facets
+- added: facet interval support
+- added: ZF2 http adapter
+- added: stats for pivot facet
+- bugfix: spellcheck 'collations' and 'correctlyspelled' updated to support Solr 5 format
+- bugfix: curl adapter now uses Solr 5 compatible headers for a GET request
+- improvement: lots of code style fixes, using the SF2 code style
+
+
 ## 3.3.0 - 2014-11-16
 
 - improvement: fixes in build.xml (use phpunit in vendor directory)